Title: 사각형 채우기 3, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240405/사각형 채우기 3/rectangle-fill-3.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int dp[1001]={0,};
+long long dp[1001]={0,};
 
 void process()
 {
